move utils folder up
