+++ examples/makeSolver.py
@@ -4,7 +4,7 @@ import re
 import yaml
 from robotmpcs.models.diff_drive_mpc_model import MpcDiffDriveModel
 from robotmpcs.models.mpcModel import MpcModel
-from robotmpcs.models.utils.utils import parse_setup
+from robotmpcs.utils.utils import parse_setup
 
 # def parse_setup(setup_file: str):
 #     with open(setup_file, "r") as setup_stream:
